+++ src/store/listeners/zoom-to.ts
@@ -2,7 +2,7 @@ import * as turf from "@turf/turf"
 import { createListenerMiddleware } from "@reduxjs/toolkit"
 import { getMap } from "@/map"
 import { SourceType } from "@/types"
-import { actions } from "../"
+import { actions } from "../actions"
 import type { RootState } from ".."
 import type { LngLatBoundsLike } from "maplibre-gl"
 import { assertUnreachable } from "@/lib"
